backend/requirements.txt
fastapi==0.111.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
python-dotenv==1.0.1
firebase-admin==6.5.0
